Add files via upload
 Image Smoother, We are given a two-dimensional grid of integers. For each cell in the grid, we need to calculate a new value based on the values of its surrounding cells. Specifically, the new value should be the floor of the average of the cell's own value and the values of its immediate neighbors (up to 8 neighbors: top-left, top, top-right, right, bottom-right, bottom, bottom-left, left).

If a cell is at the edge of the grid, it may have fewer than 8 neighbors. In this case, the average should be calculated based only on the available neighbors and the cell itself.

The goal is to create a new grid where each cell's value is replaced with this calculated average.

This abstract representation emphasizes the key elements of spatial relationship between cells in a grid and the operation of calculating a new value based on immediate neighborhood. The specifics of the image smoothing application have been abstracted away.

I found the original question very hard to read, thankfully there was a comment explaining it differently.